Minor refactor of RichText builders
- Refactored methods in RichText builders
- Updated Stockfish engine
