+++ EngineService/EngineService.csproj
@@ -49,6 +49,9 @@
     <Compile Include="UciCommands.cs" />
   </ItemGroup>
   <ItemGroup>
+    <Content Include="stockfish_15_x64_avx2.exe">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Content>
     <Content Include="stockfish_8_x64.exe">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
